requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
appnope==0.1.3
APScheduler==3.10.4
argcomplete==3.1.6
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
billiard==4.2.0
blessed==1.20.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
configparser==6.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
discord.py==2.3.2
Django==4.2.3
django-admin-logs==1.0.2
django-admin-material-dashboard==1.0.10
django-analytical==3.1.0
django-axes==6.1.1
django-chartjs==2.3.0
django-cors-headers==4.1.0
django-crispy-forms==2.1
django-csp==3.7
django-debug-toolbar==4.2.0
django-decouple==2.1
django-etc==1.4.0
django-filter==23.3
django-formtools==2.4.1
django-hitcount==1.3.5
django-import-export==3.3.3
django-jazzmin==2.6.0
django-log-viewer==1.1.7
django-mail-admin==0.3.2
django-picklefield==3.1
django-q==1.3.9
django-ratelimit==4.1.0
django-redis==5.4.0
django-request-logging==0.7.5
django-rest-framework-async==0.1.0
django-reversion==5.0.8
django-secure==1.0.2
django-simple-analytics==1.0.1
django-simple-history==3.4.0
django-sslserver==0.22
django-tracking2==0.5.1
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
escpos==2.0.0
et-xmlfile==1.1.0
executing==2.0.0
Faker==21.0.0
firebase-admin==6.3.0
Flask==3.0.0
flower==2.0.1
frozenlist==1.4.1
future==0.15.2
google-api-core==1.34.0
google-api-python-client==2.101.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.48.2
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.9.0
idna==3.4
inflection==0.5.1
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
jsonfield==3.1.0
jupyter_client==8.3.1
jupyter_core==5.3.2
kombu==5.3.4
line-bot-sdk==3.7.0
macholib==1.16.3
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.6
multidict==6.0.4
mysqlclient==2.2.0
nest-asyncio==1.5.8
numpy==1.26.2
odfpy==1.4.1
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.16.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
redis==3.5.3
requests==2.31.0
rsa==4.9
sentry-sdk==1.35.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tablib==3.5.0
tornado==6.3.3
traitlets==5.11.2
ttkbootstrap==1.10.1
ttkthemes==3.2.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.0.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.1.0
user-agents==2.2.0
viivakoodi==0.8.0
vine==5.1.0
wcwidth==0.2.8
Werkzeug==3.0.1
wrapt==1.16.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
